+++ src/forms/PermissionForm.vue
@@ -173,7 +173,7 @@
             :name="$icons.informationCircle"
           />
           <span class="permission-form__loading-error-text">
-            {{ $t('permission-form.modules-loading-error') }}
+            {{ isLoadingPermissionsErrorText }}
           </span>
         </div>
       </template>
@@ -229,6 +229,17 @@ import { storeToRefs } from 'pinia'
 import { MAX_LENGTH, MODULES } from '@/enums'
 import { requiredIf } from '@vuelidate/validators'
 import { debounce } from 'lodash-es'
+import { HTTP_STATUS_CODES } from '@distributedlab/json-api-client'
+
+type ParsedErr = {
+  originalError?: {
+    response?: {
+      status?: number
+    }
+  }
+}
+
+const NO_ACCESS_INDEX = 0
 
 const DEBOUNCE_TIMEOUT = 2000 //ms
 
@@ -256,6 +267,7 @@ const { modules } = storeToRefs(usePlatformStore())
 const { currentUserId } = useAuthStore()
 const isLoadingPermissions = ref(false)
 const isLoadingPermissionsError = ref(false)
+const loadingPermissionsResponseErrorCode = ref<HTTP_STATUS_CODES | ''>('')
 const isAccessLevelRequired = ref(false)
 const accessList = ref<ModulePermissions[]>([])
 const telegramChats = ref<TelegramChat[]>([])
@@ -276,6 +288,10 @@ const isPhoneInputDisabled = computed(() => Boolean(form.username))
 
 const parsedUsername = computed(() =>
   form.username.charAt(0) === '@' ? form.username.slice(1) : form.username,
+)
+
+const isNoAccessPermission = computed(
+  () => props.module.access_level.value === NO_ACCESS_INDEX,
 )
 
 const moduleId = computed(
@@ -306,6 +322,12 @@ const selectedTelegramChat = computed(() =>
   telegramChats.value.find(
     chat => Number(chat.id) === currentTelegramChatId.value,
   ),
+)
+
+const isLoadingPermissionsErrorText = computed(() =>
+  loadingPermissionsResponseErrorCode.value === HTTP_STATUS_CODES.CONFLICT
+    ? $t('permission-form.user-already-exist')
+    : $t('permission-form.modules-loading-error'),
 )
 
 const validationRules = computed(() => ({
@@ -376,7 +398,10 @@ const submit = async () => {
           from_user: String(currentUserId),
           to_user: String(props.id),
           payload: {
-            action: isEditForm.value ? 'update_user' : 'add_user',
+            action:
+              isEditForm.value && !isNoAccessPermission.value
+                ? 'update_user'
+                : 'add_user',
             user_id: String(props.id),
             link: form.link,
             access_level: accessLevelValue?.value,
@@ -446,6 +471,7 @@ const getAccessLevelList = async () => {
   if (!isAccessLevelCanBeChosen.value || !form.module || !form.link) return
   isLoadingPermissions.value = true
   isLoadingPermissionsError.value = false
+  loadingPermissionsResponseErrorCode.value = ''
   try {
     isAccessLevelRequired.value = false
     accessList.value = []
@@ -481,6 +507,10 @@ const getAccessLevelList = async () => {
       form.accessLevel = currentAccessLevel?.name ?? ''
     }
   } catch (e) {
+    // TODO: REWRITE ERROR HANDLING
+    loadingPermissionsResponseErrorCode.value =
+      (e as ParsedErr)?.originalError?.response?.status ||
+      HTTP_STATUS_CODES.NOT_FOUND
     isLoadingPermissionsError.value = true
     ErrorHandler.processWithoutFeedback(e)
   }